+++ android/feature/authentication/src/main/kotlin/com/hulkdx/findprofessional/feature/authentication/signup/SignUpScreen.kt
@@ -20,13 +20,12 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.hulkdx.findprofessional.core.theme.AppTheme
-import com.hulkdx.findprofessional.feature.authentication.ui.AppSnackBar
+import com.hulkdx.findprofessional.core.commonui.CUSnackBar
 import com.hulkdx.findprofessional.feature.authentication.ui.EmailTextField
 import com.hulkdx.findprofessional.feature.authentication.ui.FilledButton
 import com.hulkdx.findprofessional.feature.authentication.ui.PasswordTextField
 import com.hulkdx.findprofessional.resources.MR
 import dev.icerock.moko.resources.compose.localized
-import dev.icerock.moko.resources.desc.StringDesc
 import org.koin.androidx.compose.getViewModel
 
 @Composable
@@ -88,7 +87,7 @@ private fun SignUpScreen(
                 onClick = onSubmitClicked,
             )
         }
-        AppSnackBar(
+        CUSnackBar(
             modifier = Modifier.align(Alignment.BottomCenter),
             message = error,
             onDismiss = onErrorDismissed,
