+++ android/feature/authentication/src/main/kotlin/com/hulkdx/findprofessional/feature/authentication/ui/EmailTextField.kt
@@ -6,7 +6,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
-import com.hulkdx.findprofessional.core.commonui.CommonTextField
+import com.hulkdx.findprofessional.core.commonui.CUTextField
 import com.hulkdx.findprofessional.feature.authentication.R
 import com.hulkdx.findprofessional.resources.MR
 
@@ -16,7 +16,7 @@ fun EmailTextField(
     value: String,
     onValueChanged: (String) -> (Unit),
 ) {
-    CommonTextField(
+    CUTextField(
         modifier = modifier
             .fillMaxWidth()
             .padding(
