+++ android/feature/authentication/src/main/kotlin/com/hulkdx/findprofessional/feature/authentication/ui/PasswordTextField.kt
@@ -7,7 +7,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.input.PasswordVisualTransformation
 import androidx.compose.ui.unit.dp
-import com.hulkdx.findprofessional.core.commonui.CommonTextField
+import com.hulkdx.findprofessional.core.commonui.CUTextField
 import com.hulkdx.findprofessional.feature.authentication.R
 import com.hulkdx.findprofessional.resources.MR
 
@@ -17,7 +17,7 @@ fun PasswordTextField(
     value: String,
     onValueChanged: (String) -> (Unit),
 ) {
-    CommonTextField(
+    CUTextField(
         modifier = modifier
             .fillMaxWidth()
             .padding(
